+++ xclim/testing/tests/test_stats.py
@@ -120,18 +120,19 @@ def genextreme(request):
     return da
 
 
-def test_fit(fitda):
-    p = stats.fit(fitda, "lognorm")
+class TestFit:
+    def test_fit(self, fitda):
+        p = stats.fit(fitda, "lognorm")
 
-    assert p.dims[0] == "dparams"
-    assert p.get_axis_num("dparams") == 0
-    p0 = lognorm.fit(fitda.values[:, 0, 0])
-    np.testing.assert_array_equal(p[:, 0, 0], p0)
+        assert p.dims[0] == "dparams"
+        assert p.get_axis_num("dparams") == 0
+        p0 = lognorm.fit(fitda.values[:, 0, 0])
+        np.testing.assert_array_equal(p[:, 0, 0], p0)
 
-    # Check that we can reuse the parameters with scipy distributions
-    cdf = lognorm.cdf(0.99, *p.values)
-    assert cdf.shape == (fitda.x.size, fitda.y.size)
-    assert p.attrs["estimator"] == "Maximum likelihood"
+        # Check that we can reuse the parameters with scipy distributions
+        cdf = lognorm.cdf(0.99, *p.values)
+        assert cdf.shape == (fitda.x.size, fitda.y.size)
+        assert p.attrs["estimator"] == "Maximum likelihood"
 
 
 def test_weibull_min_fit(weibull_min):
